231027/requirements.txt
python-telegram-bot
aioschedule